requirements.txt
polars
pandas
numpy
matplotlib
networkx
community-louvain
scikit-learn
seaborn